+++ eolib/src/main/java/dev/cirras/packet/SequenceStart.java
@@ -11,4 +11,13 @@ public interface SequenceStart {
    * @return the sequence start value
    */
   int getValue();
+
+  /**
+   * Returns an instance of {@code SequenceStart} with a value of {@code 0}.
+   *
+   * @return an instance of {@code SequenceStart}
+   */
+  static SequenceStart zero() {
+    return new AbstractSequenceStart(0) {};
+  }
 }