+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `SequenceStart.zero()` method to create a `SequenceStart` instance with a value of zero.
+
 ## 1.0.0-RC1 - 2022-12-28
 
 ### Added